+++ newsroom/commands/manager.py
@@ -1,0 +1,7 @@
+from flask_script import Manager
+
+from newsroom.web.factory import get_app
+
+
+app = get_app()
+manager = Manager(app)
